Outlier Detection with a Tramo Model

Tramo is a particular regarima model estimation algorithm, mainly used to linearized the series before performing a decomposition with Seats

Usage

tramo_outliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = FALSE,
  X = NULL,
  X.td = NULL,
  ao = TRUE,
  ls = TRUE,
  tc = FALSE,
  so = FALSE,
  cv = 0,
  ml = FALSE,
  clean = FALSE
)

Arguments

y

the dependent variable (a ts object).

order, seasonal

the orders of the ARIMA model.

mean

Boolean to include or not the mean.

X

user defined regressors (other than calendar).

X.td

calendar regressors.

ao, ls, so, tc

Boolean to indicate which type of outliers should be detected.

cv

numeric. The entered critical value for the outliers' detection procedure. If equal to 0 the critical value for the outliers' detection procedure is automatically determined by the number of observations.

ml

Use of maximum likelihood (otherwise approximation by means of Hannan-Rissanen).

clean

Clean missing values at the beginning/end of the series. Regression variables are automatically resized, if need be.

Value

a "JD3_REGARIMA_OUTLIERS" object.

Examples

tramo_outliers(rjd3toolkit::ABS$X0.2.09.10.M)
#> $model
#> $model$y
#>         Jan    Feb    Mar    Apr    May    Jun    Jul    Aug    Sep    Oct
#> 1982                       460.1  502.6  443.8  459.1  438.4  465.1  452.7
#> 1983  379.2  378.0  472.1  503.4  510.6  462.4  468.3  458.2  482.7  485.3
#> 1984  414.7  414.5  484.7  487.3  597.9  500.4  543.4  503.4  522.8  556.6
#> 1985  516.3  452.5  525.8  587.7  700.3  561.8  602.8  582.5  563.1  637.1
#> 1986  570.5  478.2  547.4  594.3  751.6  553.4  663.2  581.1  661.9  665.6
#> 1987  613.9  513.2  599.9  674.1  714.0  670.5  720.9  601.6  672.3  709.1
#> 1988  631.0  551.1  678.1  715.7  740.8  722.0  683.5  650.9  723.3  729.6
#> 1989  631.5  552.0  719.0  697.6  764.8  786.3  715.1  723.8  757.9  751.7
#> 1990  678.2  586.2  726.8  744.1  815.5  832.4  710.3  759.4  741.1  786.6
#> 1991  694.0  604.7  719.2  748.2  828.2  746.9  794.5  770.4  741.5  858.6
#> 1992  740.0  665.9  701.5  831.4  878.6  826.0  788.2  723.6  819.8  902.5
#> 1993  762.1  643.0  754.1  840.7  906.6  887.1  771.5  728.7  844.7  886.9
#> 1994  745.7  664.4  821.5  831.7  908.0  912.6  782.9  798.8  887.0  934.6
#> 1995  752.4  682.5  811.2  906.0  927.2  906.8  880.6  873.9  856.8  920.6
#> 1996  833.1  737.1  812.0  895.2  962.8  908.6  908.0  888.9  833.7  933.7
#> 1997  840.9  727.4  857.9  849.0  994.8  830.2  971.1  836.0  939.1  976.9
#> 1998  917.3  716.2  822.9  970.1  970.2  849.4 1042.3  869.9  939.4 1021.3
#> 1999  942.0  738.4  903.2  953.2 1011.2  894.4 1054.5  899.5 1002.3 1043.7
#> 2000  924.9  798.2  901.9 1024.7 1052.3 1165.5  859.3 1009.2 1054.6 1070.4
#> 2001  971.9  814.6 1017.5 1039.2 1123.5 1024.9 1100.8  963.0 1012.9 1132.0
#> 2002 1027.9  841.4 1043.9 1075.3 1190.9 1143.0 1075.7 1065.9 1060.1 1211.4
#> 2003 1099.3  900.5 1092.7 1222.4 1237.1 1237.9 1182.0 1101.2 1198.2 1316.1
#> 2004 1182.9  989.8 1131.4 1277.1 1280.3 1384.1 1305.9 1166.8 1317.9 1358.3
#> 2005 1246.3 1037.3 1300.8 1153.7 1264.2 1454.2 1290.1 1210.7 1277.8 1314.4
#> 2006 1193.7 1037.7 1204.5 1348.6 1267.6 1429.0 1412.0 1239.2 1219.1 1344.6
#> 2007 1267.3 1047.0 1331.6 1302.6 1365.1 1491.5 1462.3 1315.5 1353.3 1440.6
#> 2008 1397.8 1140.5 1351.7 1396.6 1421.1 1401.6 1582.3 1268.4 1383.3 1452.4
#> 2009 1451.0 1056.6 1386.9 1509.1 1519.4 1500.5 1570.7 1341.5 1399.9 1534.3
#> 2010 1469.1 1111.9 1379.9 1389.7 1427.2 1551.4 1581.0 1324.0 1422.0 1464.9
#> 2011 1412.6 1117.5 1321.6 1472.6 1408.9 1471.9 1532.5 1293.5 1345.7 1404.7
#> 2012 1362.4 1131.7 1349.2 1391.2 1456.9 1616.4 1423.4 1359.0 1367.8 1442.6
#> 2013 1397.4 1113.6 1397.3 1339.1 1441.9 1537.4 1390.6 1337.2 1359.4 1463.3
#> 2014 1451.0 1064.9 1293.2 1442.9 1411.8 1461.6 1501.6 1254.2 1356.4 1478.7
#> 2015 1471.2 1053.8 1367.2 1442.2 1428.7 1480.9 1540.9 1331.9 1400.1 1566.3
#> 2016 1519.2 1155.8 1451.5 1451.0 1449.7 1596.1 1468.3 1293.9 1393.5 1497.4
#> 2017 1428.5 1092.4 1370.3 1522.6 1452.4 1557.2 1445.5 1303.1              
#>         Nov    Dec
#> 1982  522.9  889.3
#> 1983  568.7  963.7
#> 1984  623.2 1039.4
#> 1985  697.1 1187.5
#> 1986  700.9 1367.9
#> 1987  743.2 1460.1
#> 1988  870.3 1570.0
#> 1989  923.8 1569.4
#> 1990  931.5 1563.1
#> 1991  944.7 1600.3
#> 1992  968.6 1650.9
#> 1993  970.0 1710.5
#> 1994 1000.4 1817.5
#> 1995 1067.4 1857.2
#> 1996 1081.6 1837.6
#> 1997 1111.3 1879.1
#> 1998 1137.7 1975.7
#> 1999 1207.2 2069.6
#> 2000 1232.5 2177.5
#> 2001 1344.8 2269.5
#> 2002 1495.1 2338.6
#> 2003 1528.2 2424.2
#> 2004 1536.7 2500.8
#> 2005 1540.4 2536.0
#> 2006 1623.3 2611.1
#> 2007 1687.9 2747.0
#> 2008 1675.9 2886.1
#> 2009 1736.6 2795.1
#> 2010 1705.5 2752.4
#> 2011 1660.0 2730.5
#> 2012 1672.9 2753.3
#> 2013 1668.9 2725.5
#> 2014 1687.7 2756.9
#> 2015 1730.5 2913.6
#> 2016 1684.3 2850.4
#> 2017              
#> 
#> $model$variables
#> [1] "AO.220" "AO.219"
#> 
#> $model$X
#>        [,1] [,2]
#>   [1,]    0    0
#>   [2,]    0    0
#>   [3,]    0    0
#>   [4,]    0    0
#>   [5,]    0    0
#>   [6,]    0    0
#>   [7,]    0    0
#>   [8,]    0    0
#>   [9,]    0    0
#>  [10,]    0    0
#>  [11,]    0    0
#>  [12,]    0    0
#>  [13,]    0    0
#>  [14,]    0    0
#>  [15,]    0    0
#>  [16,]    0    0
#>  [17,]    0    0
#>  [18,]    0    0
#>  [19,]    0    0
#>  [20,]    0    0
#>  [21,]    0    0
#>  [22,]    0    0
#>  [23,]    0    0
#>  [24,]    0    0
#>  [25,]    0    0
#>  [26,]    0    0
#>  [27,]    0    0
#>  [28,]    0    0
#>  [29,]    0    0
#>  [30,]    0    0
#>  [31,]    0    0
#>  [32,]    0    0
#>  [33,]    0    0
#>  [34,]    0    0
#>  [35,]    0    0
#>  [36,]    0    0
#>  [37,]    0    0
#>  [38,]    0    0
#>  [39,]    0    0
#>  [40,]    0    0
#>  [41,]    0    0
#>  [42,]    0    0
#>  [43,]    0    0
#>  [44,]    0    0
#>  [45,]    0    0
#>  [46,]    0    0
#>  [47,]    0    0
#>  [48,]    0    0
#>  [49,]    0    0
#>  [50,]    0    0
#>  [51,]    0    0
#>  [52,]    0    0
#>  [53,]    0    0
#>  [54,]    0    0
#>  [55,]    0    0
#>  [56,]    0    0
#>  [57,]    0    0
#>  [58,]    0    0
#>  [59,]    0    0
#>  [60,]    0    0
#>  [61,]    0    0
#>  [62,]    0    0
#>  [63,]    0    0
#>  [64,]    0    0
#>  [65,]    0    0
#>  [66,]    0    0
#>  [67,]    0    0
#>  [68,]    0    0
#>  [69,]    0    0
#>  [70,]    0    0
#>  [71,]    0    0
#>  [72,]    0    0
#>  [73,]    0    0
#>  [74,]    0    0
#>  [75,]    0    0
#>  [76,]    0    0
#>  [77,]    0    0
#>  [78,]    0    0
#>  [79,]    0    0
#>  [80,]    0    0
#>  [81,]    0    0
#>  [82,]    0    0
#>  [83,]    0    0
#>  [84,]    0    0
#>  [85,]    0    0
#>  [86,]    0    0
#>  [87,]    0    0
#>  [88,]    0    0
#>  [89,]    0    0
#>  [90,]    0    0
#>  [91,]    0    0
#>  [92,]    0    0
#>  [93,]    0    0
#>  [94,]    0    0
#>  [95,]    0    0
#>  [96,]    0    0
#>  [97,]    0    0
#>  [98,]    0    0
#>  [99,]    0    0
#> [100,]    0    0
#> [101,]    0    0
#> [102,]    0    0
#> [103,]    0    0
#> [104,]    0    0
#> [105,]    0    0
#> [106,]    0    0
#> [107,]    0    0
#> [108,]    0    0
#> [109,]    0    0
#> [110,]    0    0
#> [111,]    0    0
#> [112,]    0    0
#> [113,]    0    0
#> [114,]    0    0
#> [115,]    0    0
#> [116,]    0    0
#> [117,]    0    0
#> [118,]    0    0
#> [119,]    0    0
#> [120,]    0    0
#> [121,]    0    0
#> [122,]    0    0
#> [123,]    0    0
#> [124,]    0    0
#> [125,]    0    0
#> [126,]    0    0
#> [127,]    0    0
#> [128,]    0    0
#> [129,]    0    0
#> [130,]    0    0
#> [131,]    0    0
#> [132,]    0    0
#> [133,]    0    0
#> [134,]    0    0
#> [135,]    0    0
#> [136,]    0    0
#> [137,]    0    0
#> [138,]    0    0
#> [139,]    0    0
#> [140,]    0    0
#> [141,]    0    0
#> [142,]    0    0
#> [143,]    0    0
#> [144,]    0    0
#> [145,]    0    0
#> [146,]    0    0
#> [147,]    0    0
#> [148,]    0    0
#> [149,]    0    0
#> [150,]    0    0
#> [151,]    0    0
#> [152,]    0    0
#> [153,]    0    0
#> [154,]    0    0
#> [155,]    0    0
#> [156,]    0    0
#> [157,]    0    0
#> [158,]    0    0
#> [159,]    0    0
#> [160,]    0    0
#> [161,]    0    0
#> [162,]    0    0
#> [163,]    0    0
#> [164,]    0    0
#> [165,]    0    0
#> [166,]    0    0
#> [167,]    0    0
#> [168,]    0    0
#> [169,]    0    0
#> [170,]    0    0
#> [171,]    0    0
#> [172,]    0    0
#> [173,]    0    0
#> [174,]    0    0
#> [175,]    0    0
#> [176,]    0    0
#> [177,]    0    0
#> [178,]    0    0
#> [179,]    0    0
#> [180,]    0    0
#> [181,]    0    0
#> [182,]    0    0
#> [183,]    0    0
#> [184,]    0    0
#> [185,]    0    0
#> [186,]    0    0
#> [187,]    0    0
#> [188,]    0    0
#> [189,]    0    0
#> [190,]    0    0
#> [191,]    0    0
#> [192,]    0    0
#> [193,]    0    0
#> [194,]    0    0
#> [195,]    0    0
#> [196,]    0    0
#> [197,]    0    0
#> [198,]    0    0
#> [199,]    0    0
#> [200,]    0    0
#> [201,]    0    0
#> [202,]    0    0
#> [203,]    0    0
#> [204,]    0    0
#> [205,]    0    0
#> [206,]    0    0
#> [207,]    0    0
#> [208,]    0    0
#> [209,]    0    0
#> [210,]    0    0
#> [211,]    0    0
#> [212,]    0    0
#> [213,]    0    0
#> [214,]    0    0
#> [215,]    0    0
#> [216,]    0    0
#> [217,]    0    0
#> [218,]    0    0
#> [219,]    0    1
#> [220,]    1    0
#> [221,]    0    0
#> [222,]    0    0
#> [223,]    0    0
#> [224,]    0    0
#> [225,]    0    0
#> [226,]    0    0
#> [227,]    0    0
#> [228,]    0    0
#> [229,]    0    0
#> [230,]    0    0
#> [231,]    0    0
#> [232,]    0    0
#> [233,]    0    0
#> [234,]    0    0
#> [235,]    0    0
#> [236,]    0    0
#> [237,]    0    0
#> [238,]    0    0
#> [239,]    0    0
#> [240,]    0    0
#> [241,]    0    0
#> [242,]    0    0
#> [243,]    0    0
#> [244,]    0    0
#> [245,]    0    0
#> [246,]    0    0
#> [247,]    0    0
#> [248,]    0    0
#> [249,]    0    0
#> [250,]    0    0
#> [251,]    0    0
#> [252,]    0    0
#> [253,]    0    0
#> [254,]    0    0
#> [255,]    0    0
#> [256,]    0    0
#> [257,]    0    0
#> [258,]    0    0
#> [259,]    0    0
#> [260,]    0    0
#> [261,]    0    0
#> [262,]    0    0
#> [263,]    0    0
#> [264,]    0    0
#> [265,]    0    0
#> [266,]    0    0
#> [267,]    0    0
#> [268,]    0    0
#> [269,]    0    0
#> [270,]    0    0
#> [271,]    0    0
#> [272,]    0    0
#> [273,]    0    0
#> [274,]    0    0
#> [275,]    0    0
#> [276,]    0    0
#> [277,]    0    0
#> [278,]    0    0
#> [279,]    0    0
#> [280,]    0    0
#> [281,]    0    0
#> [282,]    0    0
#> [283,]    0    0
#> [284,]    0    0
#> [285,]    0    0
#> [286,]    0    0
#> [287,]    0    0
#> [288,]    0    0
#> [289,]    0    0
#> [290,]    0    0
#> [291,]    0    0
#> [292,]    0    0
#> [293,]    0    0
#> [294,]    0    0
#> [295,]    0    0
#> [296,]    0    0
#> [297,]    0    0
#> [298,]    0    0
#> [299,]    0    0
#> [300,]    0    0
#> [301,]    0    0
#> [302,]    0    0
#> [303,]    0    0
#> [304,]    0    0
#> [305,]    0    0
#> [306,]    0    0
#> [307,]    0    0
#> [308,]    0    0
#> [309,]    0    0
#> [310,]    0    0
#> [311,]    0    0
#> [312,]    0    0
#> [313,]    0    0
#> [314,]    0    0
#> [315,]    0    0
#> [316,]    0    0
#> [317,]    0    0
#> [318,]    0    0
#> [319,]    0    0
#> [320,]    0    0
#> [321,]    0    0
#> [322,]    0    0
#> [323,]    0    0
#> [324,]    0    0
#> [325,]    0    0
#> [326,]    0    0
#> [327,]    0    0
#> [328,]    0    0
#> [329,]    0    0
#> [330,]    0    0
#> [331,]    0    0
#> [332,]    0    0
#> [333,]    0    0
#> [334,]    0    0
#> [335,]    0    0
#> [336,]    0    0
#> [337,]    0    0
#> [338,]    0    0
#> [339,]    0    0
#> [340,]    0    0
#> [341,]    0    0
#> [342,]    0    0
#> [343,]    0    0
#> [344,]    0    0
#> [345,]    0    0
#> [346,]    0    0
#> [347,]    0    0
#> [348,]    0    0
#> [349,]    0    0
#> [350,]    0    0
#> [351,]    0    0
#> [352,]    0    0
#> [353,]    0    0
#> [354,]    0    0
#> [355,]    0    0
#> [356,]    0    0
#> [357,]    0    0
#> [358,]    0    0
#> [359,]    0    0
#> [360,]    0    0
#> [361,]    0    0
#> [362,]    0    0
#> [363,]    0    0
#> [364,]    0    0
#> [365,]    0    0
#> [366,]    0    0
#> [367,]    0    0
#> [368,]    0    0
#> [369,]    0    0
#> [370,]    0    0
#> [371,]    0    0
#> [372,]    0    0
#> [373,]    0    0
#> [374,]    0    0
#> [375,]    0    0
#> [376,]    0    0
#> [377,]    0    0
#> [378,]    0    0
#> [379,]    0    0
#> [380,]    0    0
#> [381,]    0    0
#> [382,]    0    0
#> [383,]    0    0
#> [384,]    0    0
#> [385,]    0    0
#> [386,]    0    0
#> [387,]    0    0
#> [388,]    0    0
#> [389,]    0    0
#> [390,]    0    0
#> [391,]    0    0
#> [392,]    0    0
#> [393,]    0    0
#> [394,]    0    0
#> [395,]    0    0
#> [396,]    0    0
#> [397,]    0    0
#> [398,]    0    0
#> [399,]    0    0
#> [400,]    0    0
#> [401,]    0    0
#> [402,]    0    0
#> [403,]    0    0
#> [404,]    0    0
#> [405,]    0    0
#> [406,]    0    0
#> [407,]    0    0
#> [408,]    0    0
#> [409,]    0    0
#> [410,]    0    0
#> [411,]    0    0
#> [412,]    0    0
#> [413,]    0    0
#> [414,]    0    0
#> [415,]    0    0
#> [416,]    0    0
#> [417,]    0    0
#> [418,]    0    0
#> [419,]    0    0
#> [420,]    0    0
#> [421,]    0    0
#> [422,]    0    0
#> [423,]    0    0
#> [424,]    0    0
#> [425,]    0    0
#> 
#> $model$b
#> [1] -211.0035  189.6218
#> 
#> $model$bcov
#>           [,1]      [,2]
#> [1,] 1516.9191  124.4537
#> [2,]  124.4537 1516.9191
#> 
#> $model$linearized
#>   [1]  460.1000  502.6000  443.8000  459.1000  438.4000  465.1000  452.7000
#>   [8]  522.9000  889.3000  379.2000  378.0000  472.1000  503.4000  510.6000
#>  [15]  462.4000  468.3000  458.2000  482.7000  485.3000  568.7000  963.7000
#>  [22]  414.7000  414.5000  484.7000  487.3000  597.9000  500.4000  543.4000
#>  [29]  503.4000  522.8000  556.6000  623.2000 1039.4000  516.3000  452.5000
#>  [36]  525.8000  587.7000  700.3000  561.8000  602.8000  582.5000  563.1000
#>  [43]  637.1000  697.1000 1187.5000  570.5000  478.2000  547.4000  594.3000
#>  [50]  751.6000  553.4000  663.2000  581.1000  661.9000  665.6000  700.9000
#>  [57] 1367.9000  613.9000  513.2000  599.9000  674.1000  714.0000  670.5000
#>  [64]  720.9000  601.6000  672.3000  709.1000  743.2000 1460.1000  631.0000
#>  [71]  551.1000  678.1000  715.7000  740.8000  722.0000  683.5000  650.9000
#>  [78]  723.3000  729.6000  870.3000 1570.0000  631.5000  552.0000  719.0000
#>  [85]  697.6000  764.8000  786.3000  715.1000  723.8000  757.9000  751.7000
#>  [92]  923.8000 1569.4000  678.2000  586.2000  726.8000  744.1000  815.5000
#>  [99]  832.4000  710.3000  759.4000  741.1000  786.6000  931.5000 1563.1000
#> [106]  694.0000  604.7000  719.2000  748.2000  828.2000  746.9000  794.5000
#> [113]  770.4000  741.5000  858.6000  944.7000 1600.3000  740.0000  665.9000
#> [120]  701.5000  831.4000  878.6000  826.0000  788.2000  723.6000  819.8000
#> [127]  902.5000  968.6000 1650.9000  762.1000  643.0000  754.1000  840.7000
#> [134]  906.6000  887.1000  771.5000  728.7000  844.7000  886.9000  970.0000
#> [141] 1710.5000  745.7000  664.4000  821.5000  831.7000  908.0000  912.6000
#> [148]  782.9000  798.8000  887.0000  934.6000 1000.4000 1817.5000  752.4000
#> [155]  682.5000  811.2000  906.0000  927.2000  906.8000  880.6000  873.9000
#> [162]  856.8000  920.6000 1067.4000 1857.2000  833.1000  737.1000  812.0000
#> [169]  895.2000  962.8000  908.6000  908.0000  888.9000  833.7000  933.7000
#> [176] 1081.6000 1837.6000  840.9000  727.4000  857.9000  849.0000  994.8000
#> [183]  830.2000  971.1000  836.0000  939.1000  976.9000 1111.3000 1879.1000
#> [190]  917.3000  716.2000  822.9000  970.1000  970.2000  849.4000 1042.3000
#> [197]  869.9000  939.4000 1021.3000 1137.7000 1975.7000  942.0000  738.4000
#> [204]  903.2000  953.2000 1011.2000  894.4000 1054.5000  899.5000 1002.3000
#> [211] 1043.7000 1207.2000 2069.6000  924.9000  798.2000  901.9000 1024.7000
#> [218] 1052.3000  975.8782 1070.3035 1009.2000 1054.6000 1070.4000 1232.5000
#> [225] 2177.5000  971.9000  814.6000 1017.5000 1039.2000 1123.5000 1024.9000
#> [232] 1100.8000  963.0000 1012.9000 1132.0000 1344.8000 2269.5000 1027.9000
#> [239]  841.4000 1043.9000 1075.3000 1190.9000 1143.0000 1075.7000 1065.9000
#> [246] 1060.1000 1211.4000 1495.1000 2338.6000 1099.3000  900.5000 1092.7000
#> [253] 1222.4000 1237.1000 1237.9000 1182.0000 1101.2000 1198.2000 1316.1000
#> [260] 1528.2000 2424.2000 1182.9000  989.8000 1131.4000 1277.1000 1280.3000
#> [267] 1384.1000 1305.9000 1166.8000 1317.9000 1358.3000 1536.7000 2500.8000
#> [274] 1246.3000 1037.3000 1300.8000 1153.7000 1264.2000 1454.2000 1290.1000
#> [281] 1210.7000 1277.8000 1314.4000 1540.4000 2536.0000 1193.7000 1037.7000
#> [288] 1204.5000 1348.6000 1267.6000 1429.0000 1412.0000 1239.2000 1219.1000
#> [295] 1344.6000 1623.3000 2611.1000 1267.3000 1047.0000 1331.6000 1302.6000
#> [302] 1365.1000 1491.5000 1462.3000 1315.5000 1353.3000 1440.6000 1687.9000
#> [309] 2747.0000 1397.8000 1140.5000 1351.7000 1396.6000 1421.1000 1401.6000
#> [316] 1582.3000 1268.4000 1383.3000 1452.4000 1675.9000 2886.1000 1451.0000
#> [323] 1056.6000 1386.9000 1509.1000 1519.4000 1500.5000 1570.7000 1341.5000
#> [330] 1399.9000 1534.3000 1736.6000 2795.1000 1469.1000 1111.9000 1379.9000
#> [337] 1389.7000 1427.2000 1551.4000 1581.0000 1324.0000 1422.0000 1464.9000
#> [344] 1705.5000 2752.4000 1412.6000 1117.5000 1321.6000 1472.6000 1408.9000
#> [351] 1471.9000 1532.5000 1293.5000 1345.7000 1404.7000 1660.0000 2730.5000
#> [358] 1362.4000 1131.7000 1349.2000 1391.2000 1456.9000 1616.4000 1423.4000
#> [365] 1359.0000 1367.8000 1442.6000 1672.9000 2753.3000 1397.4000 1113.6000
#> [372] 1397.3000 1339.1000 1441.9000 1537.4000 1390.6000 1337.2000 1359.4000
#> [379] 1463.3000 1668.9000 2725.5000 1451.0000 1064.9000 1293.2000 1442.9000
#> [386] 1411.8000 1461.6000 1501.6000 1254.2000 1356.4000 1478.7000 1687.7000
#> [393] 2756.9000 1471.2000 1053.8000 1367.2000 1442.2000 1428.7000 1480.9000
#> [400] 1540.9000 1331.9000 1400.1000 1566.3000 1730.5000 2913.6000 1519.2000
#> [407] 1155.8000 1451.5000 1451.0000 1449.7000 1596.1000 1468.3000 1293.9000
#> [414] 1393.5000 1497.4000 1684.3000 2850.4000 1428.5000 1092.4000 1370.3000
#> [421] 1522.6000 1452.4000 1557.2000 1445.5000 1303.1000
#> 
#> 
#> $likelihood
#> $likelihood$initial
#> $likelihood$initial$ll
#> [1] -2218.964
#> 
#> $likelihood$initial$ssq
#> [1] 1139531
#> 
#> $likelihood$initial$nobs
#> [1] 425
#> 
#> $likelihood$initial$neffective
#> [1] -1
#> 
#> $likelihood$initial$nparams
#> [1] 3
#> 
#> $likelihood$initial$df
#> [1] 409
#> 
#> $likelihood$initial$aic
#> [1] 4443.928
#> 
#> $likelihood$initial$aicc
#> [1] 4443.987
#> 
#> $likelihood$initial$bic
#> [1] 4455.991
#> 
#> $likelihood$initial$bic2
#> [1] 10.81551
#> 
#> $likelihood$initial$bicc
#> [1] 7.954332
#> 
#> $likelihood$initial$hannanquinn
#> [1] 4448.7
#> 
#> 
#> $likelihood$final
#> $likelihood$final$ll
#> [1] -2194.499
#> 
#> $likelihood$final$ssq
#> [1] 1014171
#> 
#> $likelihood$final$nobs
#> [1] 425
#> 
#> $likelihood$final$neffective
#> [1] -1
#> 
#> $likelihood$final$nparams
#> [1] 3
#> 
#> $likelihood$final$df
#> [1] 409
#> 
#> $likelihood$final$aic
#> [1] 4394.998
#> 
#> $likelihood$final$aicc
#> [1] 4395.057
#> 
#> $likelihood$final$bic
#> [1] 4407.061
#> 
#> $likelihood$final$bic2
#> [1] 10.69675
#> 
#> $likelihood$final$bicc
#> [1] 7.837787
#> 
#> $likelihood$final$hannanquinn
#> [1] 4399.77
#> 
#> 
#> 
#> attr(,"class")
#> [1] "JD3_REGARIMA_OUTLIERS"