Display list of all available output objects (series, parameters, diagnostics)

Function generating a comprehensive list of available output variables from the estimation process by the tramo() and tramoseats() functions. Some items are available in the default estimation output but the remainder can be added using the userdefined parameter. User-defined objects can the be retrieved from the list of lists generated by the estimation process

Usage

userdefined_variables_tramoseats(x = c("TRAMO-SEATS", "TRAMO"))

Arguments

x

a character to indicate the estimation function for which the output items list will be displayed.

Value

a vector containing the names of all the available output objects (series, diagnostics, parameters)

References

More information and examples related to 'JDemetra+' features in the online documentation: https://jdemetra-new-documentation.netlify.app/

Examples

userdefined_variables_tramoseats("tramoseats")
#>   [1] "adjust"                              
#>   [2] "arima.bd"                            
#>   [3] "arima.bp"                            
#>   [4] "arima.bphi(*)"                       
#>   [5] "arima.bq"                            
#>   [6] "arima.btheta(*)"                     
#>   [7] "arima.d"                             
#>   [8] "arima.p"                             
#>   [9] "arima.phi(*)"                        
#>  [10] "arima.q"                             
#>  [11] "arima.theta(*)"                      
#>  [12] "benchmarking.original"               
#>  [13] "benchmarking.result"                 
#>  [14] "benchmarking.target"                 
#>  [15] "cal"                                 
#>  [16] "cal_b"                               
#>  [17] "cal_b(?)"                            
#>  [18] "cal_f"                               
#>  [19] "cal_f(?)"                            
#>  [20] "decomposition.finalmodel.ac"         
#>  [21] "decomposition.finalmodel.bd"         
#>  [22] "decomposition.finalmodel.bp"         
#>  [23] "decomposition.finalmodel.bphi(*)"    
#>  [24] "decomposition.finalmodel.bq"         
#>  [25] "decomposition.finalmodel.btheta(*)"  
#>  [26] "decomposition.finalmodel.d"          
#>  [27] "decomposition.finalmodel.p"          
#>  [28] "decomposition.finalmodel.phi(*)"     
#>  [29] "decomposition.finalmodel.q"          
#>  [30] "decomposition.finalmodel.spectrum"   
#>  [31] "decomposition.finalmodel.theta(*)"   
#>  [32] "decomposition.i_cmp"                 
#>  [33] "decomposition.i_cmp_b"               
#>  [34] "decomposition.i_cmp_eb"              
#>  [35] "decomposition.i_cmp_ef"              
#>  [36] "decomposition.i_cmp_f"               
#>  [37] "decomposition.i_lin"                 
#>  [38] "decomposition.i_lin_b"               
#>  [39] "decomposition.i_lin_e"               
#>  [40] "decomposition.i_lin_eb"              
#>  [41] "decomposition.i_lin_ef"              
#>  [42] "decomposition.i_lin_f"               
#>  [43] "decomposition.initialmodel.ac"       
#>  [44] "decomposition.initialmodel.bd"       
#>  [45] "decomposition.initialmodel.bp"       
#>  [46] "decomposition.initialmodel.bphi(*)"  
#>  [47] "decomposition.initialmodel.bq"       
#>  [48] "decomposition.initialmodel.btheta(*)"
#>  [49] "decomposition.initialmodel.d"        
#>  [50] "decomposition.initialmodel.p"        
#>  [51] "decomposition.initialmodel.phi(*)"   
#>  [52] "decomposition.initialmodel.q"        
#>  [53] "decomposition.initialmodel.spectrum" 
#>  [54] "decomposition.initialmodel.theta(*)" 
#>  [55] "decomposition.s_cmp"                 
#>  [56] "decomposition.s_cmp_b"               
#>  [57] "decomposition.s_cmp_eb"              
#>  [58] "decomposition.s_cmp_ef"              
#>  [59] "decomposition.s_cmp_f"               
#>  [60] "decomposition.s_lin"                 
#>  [61] "decomposition.s_lin_b"               
#>  [62] "decomposition.s_lin_e"               
#>  [63] "decomposition.s_lin_eb"              
#>  [64] "decomposition.s_lin_ef"              
#>  [65] "decomposition.s_lin_f"               
#>  [66] "decomposition.sa_cmp"                
#>  [67] "decomposition.sa_cmp_b"              
#>  [68] "decomposition.sa_cmp_eb"             
#>  [69] "decomposition.sa_cmp_ef"             
#>  [70] "decomposition.sa_cmp_f"              
#>  [71] "decomposition.sa_lin"                
#>  [72] "decomposition.sa_lin_b"              
#>  [73] "decomposition.sa_lin_e"              
#>  [74] "decomposition.sa_lin_eb"             
#>  [75] "decomposition.sa_lin_ef"             
#>  [76] "decomposition.sa_lin_f"              
#>  [77] "decomposition.t_cmp"                 
#>  [78] "decomposition.t_cmp_b"               
#>  [79] "decomposition.t_cmp_eb"              
#>  [80] "decomposition.t_cmp_ef"              
#>  [81] "decomposition.t_cmp_f"               
#>  [82] "decomposition.t_lin"                 
#>  [83] "decomposition.t_lin_b"               
#>  [84] "decomposition.t_lin_e"               
#>  [85] "decomposition.t_lin_eb"              
#>  [86] "decomposition.t_lin_ef"              
#>  [87] "decomposition.t_lin_f"               
#>  [88] "decomposition.y_cmp"                 
#>  [89] "decomposition.y_cmp_b"               
#>  [90] "decomposition.y_cmp_eb"              
#>  [91] "decomposition.y_cmp_ef"              
#>  [92] "decomposition.y_cmp_f"               
#>  [93] "decomposition.y_lin"                 
#>  [94] "decomposition.y_lin_b"               
#>  [95] "decomposition.y_lin_eb"              
#>  [96] "decomposition.y_lin_ef"              
#>  [97] "decomposition.y_lin_f"               
#>  [98] "det"                                 
#>  [99] "det_b(?)"                            
#> [100] "det_f(?)"                            
#> [101] "det_i"                               
#> [102] "det_i_b(?)"                          
#> [103] "det_i_f(?)"                          
#> [104] "det_s"                               
#> [105] "det_s_b(?)"                          
#> [106] "det_s_f(?)"                          
#> [107] "det_t"                               
#> [108] "det_t_b(?)"                          
#> [109] "det_t_f(?)"                          
#> [110] "diagnostics.fcast-insample-mean"     
#> [111] "diagnostics.fcast-outsample-mean"    
#> [112] "diagnostics.fcast-outsample-variance"
#> [113] "diagnostics.seas-i-combined"         
#> [114] "diagnostics.seas-i-combined3"        
#> [115] "diagnostics.seas-i-evolutive"        
#> [116] "diagnostics.seas-i-f"                
#> [117] "diagnostics.seas-i-friedman"         
#> [118] "diagnostics.seas-i-kw"               
#> [119] "diagnostics.seas-i-periodogram"      
#> [120] "diagnostics.seas-i-qs"               
#> [121] "diagnostics.seas-i-spectralpeaks"    
#> [122] "diagnostics.seas-i-stable"           
#> [123] "diagnostics.seas-lin-combined"       
#> [124] "diagnostics.seas-lin-evolutive"      
#> [125] "diagnostics.seas-lin-f"              
#> [126] "diagnostics.seas-lin-friedman"       
#> [127] "diagnostics.seas-lin-kw"             
#> [128] "diagnostics.seas-lin-periodogram"    
#> [129] "diagnostics.seas-lin-qs"             
#> [130] "diagnostics.seas-lin-spectralpeaks"  
#> [131] "diagnostics.seas-lin-stable"         
#> [132] "diagnostics.seas-res-combined"       
#> [133] "diagnostics.seas-res-combined3"      
#> [134] "diagnostics.seas-res-evolutive"      
#> [135] "diagnostics.seas-res-f"              
#> [136] "diagnostics.seas-res-friedman"       
#> [137] "diagnostics.seas-res-kw"             
#> [138] "diagnostics.seas-res-periodogram"    
#> [139] "diagnostics.seas-res-qs"             
#> [140] "diagnostics.seas-res-spectralpeaks"  
#> [141] "diagnostics.seas-res-stable"         
#> [142] "diagnostics.seas-sa-ac1"             
#> [143] "diagnostics.seas-sa-combined"        
#> [144] "diagnostics.seas-sa-combined3"       
#> [145] "diagnostics.seas-sa-evolutive"       
#> [146] "diagnostics.seas-sa-f"               
#> [147] "diagnostics.seas-sa-friedman"        
#> [148] "diagnostics.seas-sa-kw"              
#> [149] "diagnostics.seas-sa-periodogram"     
#> [150] "diagnostics.seas-sa-qs"              
#> [151] "diagnostics.seas-sa-spectralpeaks"   
#> [152] "diagnostics.seas-sa-stable"          
#> [153] "diagnostics.seas-si-combined"        
#> [154] "diagnostics.seas-si-combined3"       
#> [155] "diagnostics.seas-si-evolutive"       
#> [156] "diagnostics.seas-si-stable"          
#> [157] "diagnostics.td-i-all"                
#> [158] "diagnostics.td-i-last"               
#> [159] "diagnostics.td-res-all"              
#> [160] "diagnostics.td-res-last"             
#> [161] "diagnostics.td-sa-all"               
#> [162] "diagnostics.td-sa-last"              
#> [163] "ee"                                  
#> [164] "ee_b(?)"                             
#> [165] "ee_f(?)"                             
#> [166] "i"                                   
#> [167] "i_b"                                 
#> [168] "i_eb"                                
#> [169] "i_ef"                                
#> [170] "i_f"                                 
#> [171] "l"                                   
#> [172] "likelihood.adjustedll"               
#> [173] "likelihood.aic"                      
#> [174] "likelihood.aicc"                     
#> [175] "likelihood.bic"                      
#> [176] "likelihood.bic2"                     
#> [177] "likelihood.bicc"                     
#> [178] "likelihood.df"                       
#> [179] "likelihood.hannanquinn"              
#> [180] "likelihood.ll"                       
#> [181] "likelihood.neffectiveobs"            
#> [182] "likelihood.nobs"                     
#> [183] "likelihood.nparams"                  
#> [184] "likelihood.ssqerr"                   
#> [185] "log"                                 
#> [186] "mhe"                                 
#> [187] "mhe_b(?)"                            
#> [188] "mhe_f(?)"                            
#> [189] "mode"                                
#> [190] "omhe"                                
#> [191] "omhe_b(?)"                           
#> [192] "omhe_f(?)"                           
#> [193] "out"                                 
#> [194] "out_b(?)"                            
#> [195] "out_f(?)"                            
#> [196] "out_i"                               
#> [197] "out_i_b(?)"                          
#> [198] "out_i_f(?)"                          
#> [199] "out_s"                               
#> [200] "out_s_b(?)"                          
#> [201] "out_s_f(?)"                          
#> [202] "out_t"                               
#> [203] "out_t_b(?)"                          
#> [204] "out_t_f(?)"                          
#> [205] "period"                              
#> [206] "reg_i"                               
#> [207] "reg_i_b(?)"                          
#> [208] "reg_i_f(?)"                          
#> [209] "reg_s"                               
#> [210] "reg_s_b(?)"                          
#> [211] "reg_s_f(?)"                          
#> [212] "reg_sa"                              
#> [213] "reg_sa_b(?)"                         
#> [214] "reg_sa_f(?)"                         
#> [215] "reg_t"                               
#> [216] "reg_t_b(?)"                          
#> [217] "reg_t_f(?)"                          
#> [218] "reg_u"                               
#> [219] "reg_u_b(?)"                          
#> [220] "reg_u_f(?)"                          
#> [221] "reg_y"                               
#> [222] "reg_y_b(?)"                          
#> [223] "reg_y_f(?)"                          
#> [224] "regression.description"              
#> [225] "regression.details.coefficients"     
#> [226] "regression.details.covar"            
#> [227] "regression.details.covar-ml"         
#> [228] "regression.easter"                   
#> [229] "regression.espan.end"                
#> [230] "regression.espan.missing"            
#> [231] "regression.espan.n"                  
#> [232] "regression.espan.start"              
#> [233] "regression.leaster"                  
#> [234] "regression.lp"                       
#> [235] "regression.mean"                     
#> [236] "regression.missing(*)"               
#> [237] "regression.ml.parameters"            
#> [238] "regression.ml.pcorr"                 
#> [239] "regression.ml.pcovar"                
#> [240] "regression.ml.pcovar-ml"             
#> [241] "regression.ml.pscore"                
#> [242] "regression.mu"                       
#> [243] "regression.nao"                      
#> [244] "regression.nlp"                      
#> [245] "regression.nls"                      
#> [246] "regression.nmh"                      
#> [247] "regression.nout"                     
#> [248] "regression.nso"                      
#> [249] "regression.ntc"                      
#> [250] "regression.ntd"                      
#> [251] "regression.nusers"                   
#> [252] "regression.out(*)"                   
#> [253] "regression.outlier(*)"               
#> [254] "regression.td(*)"                    
#> [255] "regression.type"                     
#> [256] "regression.user(*)"                  
#> [257] "residuals.bp"                        
#> [258] "residuals.bp2"                       
#> [259] "residuals.doornikhansen"             
#> [260] "residuals.kurtosis"                  
#> [261] "residuals.lb"                        
#> [262] "residuals.lb2"                       
#> [263] "residuals.lruns"                     
#> [264] "residuals.ludruns"                   
#> [265] "residuals.mean"                      
#> [266] "residuals.nruns"                     
#> [267] "residuals.nudruns"                   
#> [268] "residuals.res"                       
#> [269] "residuals.seasbp"                    
#> [270] "residuals.seaslb"                    
#> [271] "residuals.ser"                       
#> [272] "residuals.skewness"                  
#> [273] "residuals.tsres"                     
#> [274] "residuals.type"                      
#> [275] "s"                                   
#> [276] "s_b"                                 
#> [277] "s_eb"                                
#> [278] "s_ef"                                
#> [279] "s_f"                                 
#> [280] "sa"                                  
#> [281] "sa_b"                                
#> [282] "sa_eb"                               
#> [283] "sa_ef"                               
#> [284] "sa_f"                                
#> [285] "seasonal"                            
#> [286] "seats.ar_root(*)"                    
#> [287] "seats.ivar-estimate"                 
#> [288] "seats.ivar-estimator"                
#> [289] "seats.ivar-pvalue"                   
#> [290] "seats.ma_root(*)"                    
#> [291] "seats.model_changed"                 
#> [292] "seats.parameters_cutoff"             
#> [293] "seats.savar-estimate"                
#> [294] "seats.savar-estimator"               
#> [295] "seats.savar-pvalue"                  
#> [296] "seats.seasonality"                   
#> [297] "seats.sicorr-estimate"               
#> [298] "seats.sicorr-estimator"              
#> [299] "seats.sicorr-pvalue"                 
#> [300] "seats.svar-estimate"                 
#> [301] "seats.svar-estimator"                
#> [302] "seats.svar-pvalue"                   
#> [303] "seats.ticorr-estimate"               
#> [304] "seats.ticorr-estimator"              
#> [305] "seats.ticorr-pvalue"                 
#> [306] "seats.tscorr-estimate"               
#> [307] "seats.tscorr-estimator"              
#> [308] "seats.tscorr-pvalue"                 
#> [309] "seats.tvar-estimate"                 
#> [310] "seats.tvar-estimator"                
#> [311] "seats.tvar-pvalue"                   
#> [312] "span.end"                            
#> [313] "span.missing"                        
#> [314] "span.n"                              
#> [315] "span.start"                          
#> [316] "t"                                   
#> [317] "t_b"                                 
#> [318] "t_eb"                                
#> [319] "t_ef"                                
#> [320] "t_f"                                 
#> [321] "tde"                                 
#> [322] "tde_b(?)"                            
#> [323] "tde_f(?)"                            
#> [324] "variancedecomposition.cycle"         
#> [325] "variancedecomposition.irregular"     
#> [326] "variancedecomposition.others"        
#> [327] "variancedecomposition.seasonality"   
#> [328] "variancedecomposition.tdh"           
#> [329] "variancedecomposition.total"         
#> [330] "y"                                   
#> [331] "y_b"                                 
#> [332] "y_b(?)"                              
#> [333] "y_eb(?)"                             
#> [334] "y_ef(?)"                             
#> [335] "y_f"                                 
#> [336] "y_f(?)"                              
#> [337] "yc"                                  
#> [338] "ycal"                                
#> [339] "ycal_f(?)"                           
userdefined_variables_tramoseats("tramo")
#>   [1] "adjust"                          "arima.bd"                       
#>   [3] "arima.bp"                        "arima.bphi(*)"                  
#>   [5] "arima.bq"                        "arima.btheta(*)"                
#>   [7] "arima.d"                         "arima.p"                        
#>   [9] "arima.phi(*)"                    "arima.q"                        
#>  [11] "arima.theta(*)"                  "cal"                            
#>  [13] "cal_b(?)"                        "cal_f(?)"                       
#>  [15] "det"                             "det_b(?)"                       
#>  [17] "det_f(?)"                        "det_i"                          
#>  [19] "det_i_b(?)"                      "det_i_f(?)"                     
#>  [21] "det_s"                           "det_s_b(?)"                     
#>  [23] "det_s_f(?)"                      "det_t"                          
#>  [25] "det_t_b(?)"                      "det_t_f(?)"                     
#>  [27] "ee"                              "ee_b(?)"                        
#>  [29] "ee_f(?)"                         "l"                              
#>  [31] "likelihood.adjustedll"           "likelihood.aic"                 
#>  [33] "likelihood.aicc"                 "likelihood.bic"                 
#>  [35] "likelihood.bic2"                 "likelihood.bicc"                
#>  [37] "likelihood.df"                   "likelihood.hannanquinn"         
#>  [39] "likelihood.ll"                   "likelihood.neffectiveobs"       
#>  [41] "likelihood.nobs"                 "likelihood.nparams"             
#>  [43] "likelihood.ssqerr"               "log"                            
#>  [45] "mhe"                             "mhe_b(?)"                       
#>  [47] "mhe_f(?)"                        "omhe"                           
#>  [49] "omhe_b(?)"                       "omhe_f(?)"                      
#>  [51] "out"                             "out_b(?)"                       
#>  [53] "out_f(?)"                        "out_i"                          
#>  [55] "out_i_b(?)"                      "out_i_f(?)"                     
#>  [57] "out_s"                           "out_s_b(?)"                     
#>  [59] "out_s_f(?)"                      "out_t"                          
#>  [61] "out_t_b(?)"                      "out_t_f(?)"                     
#>  [63] "period"                          "reg_i"                          
#>  [65] "reg_i_b(?)"                      "reg_i_f(?)"                     
#>  [67] "reg_s"                           "reg_s_b(?)"                     
#>  [69] "reg_s_f(?)"                      "reg_sa"                         
#>  [71] "reg_sa_b(?)"                     "reg_sa_f(?)"                    
#>  [73] "reg_t"                           "reg_t_b(?)"                     
#>  [75] "reg_t_f(?)"                      "reg_u"                          
#>  [77] "reg_u_b(?)"                      "reg_u_f(?)"                     
#>  [79] "reg_y"                           "reg_y_b(?)"                     
#>  [81] "reg_y_f(?)"                      "regression.description"         
#>  [83] "regression.details.coefficients" "regression.details.covar"       
#>  [85] "regression.details.covar-ml"     "regression.easter"              
#>  [87] "regression.espan.end"            "regression.espan.missing"       
#>  [89] "regression.espan.n"              "regression.espan.start"         
#>  [91] "regression.leaster"              "regression.lp"                  
#>  [93] "regression.mean"                 "regression.missing(*)"          
#>  [95] "regression.ml.parameters"        "regression.ml.pcorr"            
#>  [97] "regression.ml.pcovar"            "regression.ml.pcovar-ml"        
#>  [99] "regression.ml.pscore"            "regression.mu"                  
#> [101] "regression.nao"                  "regression.nlp"                 
#> [103] "regression.nls"                  "regression.nmh"                 
#> [105] "regression.nout"                 "regression.nso"                 
#> [107] "regression.ntc"                  "regression.ntd"                 
#> [109] "regression.nusers"               "regression.out(*)"              
#> [111] "regression.outlier(*)"           "regression.td(*)"               
#> [113] "regression.type"                 "regression.user(*)"             
#> [115] "residuals.bp"                    "residuals.bp2"                  
#> [117] "residuals.doornikhansen"         "residuals.kurtosis"             
#> [119] "residuals.lb"                    "residuals.lb2"                  
#> [121] "residuals.lruns"                 "residuals.ludruns"              
#> [123] "residuals.mean"                  "residuals.nruns"                
#> [125] "residuals.nudruns"               "residuals.res"                  
#> [127] "residuals.seasbp"                "residuals.seaslb"               
#> [129] "residuals.ser"                   "residuals.skewness"             
#> [131] "residuals.tsres"                 "residuals.type"                 
#> [133] "span.end"                        "span.missing"                   
#> [135] "span.n"                          "span.start"                     
#> [137] "tde"                             "tde_b(?)"                       
#> [139] "tde_f(?)"                        "y"                              
#> [141] "y_b(?)"                          "y_eb(?)"                        
#> [143] "y_ef(?)"                         "y_f(?)"                         
#> [145] "yc"                              "ycal"                           
#> [147] "ycal_f(?)"                      
y = rjd3toolkit::ABS$X0.2.09.10.M
m <- tramoseats(y,"rsafull", userdefined=c("ycal","variancedecomposition.seasonality"))
#> Error in .jcall("jdplus/tramoseats/base/r/TramoSeats", "[B", "toBuffer",     jq): java.lang.NoSuchMethodError: 'jdplus.toolkit.base.protobuf.regarima.RegArimaProtos$RegArimaModel jdplus.toolkit.base.protobuf.regarima.RegArimaEstimationProto.convert(jdplus.toolkit.base.core.regsarima.regular.RegSarimaModel)'
m$user_defined$ycal
#> Error in eval(expr, envir, enclos): object 'm' not found
m$user_defined$variancedecomposition.seasonality
#> Error in eval(expr, envir, enclos): object 'm' not found