Display list of all available output objects (series, parameters, diagnostics)

Function generating a comprehensive list of available output variables from the estimation process by the tramo() and tramoseats() functions. Some items are available in the default estimation output but the remainder can be added using the userdefined parameter. User-defined objects can the be retrieved from the list of lists generated by the estimation process

Usage

userdefined_variables_tramoseats(x = c("TRAMO-SEATS", "TRAMO"))

Arguments

x

a character to indicate the estimation function for which the output items list will be displayed.

Value

a vector containing the names of all the available output objects (series, diagnostics, parameters)

References

More information and examples related to 'JDemetra+' features in the online documentation: https://jdemetra-new-documentation.netlify.app/

Examples

userdefined_variables_tramoseats("tramoseats")
#>   [1] "adjust"                              
#>   [2] "arima.bd"                            
#>   [3] "arima.bp"                            
#>   [4] "arima.bphi(*)"                       
#>   [5] "arima.bq"                            
#>   [6] "arima.btheta(*)"                     
#>   [7] "arima.d"                             
#>   [8] "arima.p"                             
#>   [9] "arima.phi(*)"                        
#>  [10] "arima.q"                             
#>  [11] "arima.theta(*)"                      
#>  [12] "benchmarking.original"               
#>  [13] "benchmarking.result"                 
#>  [14] "benchmarking.target"                 
#>  [15] "cal"                                 
#>  [16] "cal_b"                               
#>  [17] "cal_b(?)"                            
#>  [18] "cal_f"                               
#>  [19] "cal_f(?)"                            
#>  [20] "decomposition.finalmodel.ac"         
#>  [21] "decomposition.finalmodel.bd"         
#>  [22] "decomposition.finalmodel.bp"         
#>  [23] "decomposition.finalmodel.bphi(*)"    
#>  [24] "decomposition.finalmodel.bq"         
#>  [25] "decomposition.finalmodel.btheta(*)"  
#>  [26] "decomposition.finalmodel.d"          
#>  [27] "decomposition.finalmodel.p"          
#>  [28] "decomposition.finalmodel.phi(*)"     
#>  [29] "decomposition.finalmodel.q"          
#>  [30] "decomposition.finalmodel.spectrum"   
#>  [31] "decomposition.finalmodel.theta(*)"   
#>  [32] "decomposition.i_cmp"                 
#>  [33] "decomposition.i_cmp_b"               
#>  [34] "decomposition.i_cmp_eb"              
#>  [35] "decomposition.i_cmp_ef"              
#>  [36] "decomposition.i_cmp_f"               
#>  [37] "decomposition.i_lin"                 
#>  [38] "decomposition.i_lin_b"               
#>  [39] "decomposition.i_lin_e"               
#>  [40] "decomposition.i_lin_eb"              
#>  [41] "decomposition.i_lin_ef"              
#>  [42] "decomposition.i_lin_f"               
#>  [43] "decomposition.initialmodel.ac"       
#>  [44] "decomposition.initialmodel.bd"       
#>  [45] "decomposition.initialmodel.bp"       
#>  [46] "decomposition.initialmodel.bphi(*)"  
#>  [47] "decomposition.initialmodel.bq"       
#>  [48] "decomposition.initialmodel.btheta(*)"
#>  [49] "decomposition.initialmodel.d"        
#>  [50] "decomposition.initialmodel.p"        
#>  [51] "decomposition.initialmodel.phi(*)"   
#>  [52] "decomposition.initialmodel.q"        
#>  [53] "decomposition.initialmodel.spectrum" 
#>  [54] "decomposition.initialmodel.theta(*)" 
#>  [55] "decomposition.s_cmp"                 
#>  [56] "decomposition.s_cmp_b"               
#>  [57] "decomposition.s_cmp_eb"              
#>  [58] "decomposition.s_cmp_ef"              
#>  [59] "decomposition.s_cmp_f"               
#>  [60] "decomposition.s_lin"                 
#>  [61] "decomposition.s_lin_b"               
#>  [62] "decomposition.s_lin_e"               
#>  [63] "decomposition.s_lin_eb"              
#>  [64] "decomposition.s_lin_ef"              
#>  [65] "decomposition.s_lin_f"               
#>  [66] "decomposition.sa_cmp"                
#>  [67] "decomposition.sa_cmp_b"              
#>  [68] "decomposition.sa_cmp_eb"             
#>  [69] "decomposition.sa_cmp_ef"             
#>  [70] "decomposition.sa_cmp_f"              
#>  [71] "decomposition.sa_lin"                
#>  [72] "decomposition.sa_lin_b"              
#>  [73] "decomposition.sa_lin_e"              
#>  [74] "decomposition.sa_lin_eb"             
#>  [75] "decomposition.sa_lin_ef"             
#>  [76] "decomposition.sa_lin_f"              
#>  [77] "decomposition.t_cmp"                 
#>  [78] "decomposition.t_cmp_b"               
#>  [79] "decomposition.t_cmp_eb"              
#>  [80] "decomposition.t_cmp_ef"              
#>  [81] "decomposition.t_cmp_f"               
#>  [82] "decomposition.t_lin"                 
#>  [83] "decomposition.t_lin_b"               
#>  [84] "decomposition.t_lin_e"               
#>  [85] "decomposition.t_lin_eb"              
#>  [86] "decomposition.t_lin_ef"              
#>  [87] "decomposition.t_lin_f"               
#>  [88] "decomposition.y_cmp"                 
#>  [89] "decomposition.y_cmp_b"               
#>  [90] "decomposition.y_cmp_eb"              
#>  [91] "decomposition.y_cmp_ef"              
#>  [92] "decomposition.y_cmp_f"               
#>  [93] "decomposition.y_lin"                 
#>  [94] "decomposition.y_lin_b"               
#>  [95] "decomposition.y_lin_eb"              
#>  [96] "decomposition.y_lin_ef"              
#>  [97] "decomposition.y_lin_f"               
#>  [98] "det"                                 
#>  [99] "det_b(?)"                            
#> [100] "det_f(?)"                            
#> [101] "det_i"                               
#> [102] "det_i_b(?)"                          
#> [103] "det_i_f(?)"                          
#> [104] "det_s"                               
#> [105] "det_s_b(?)"                          
#> [106] "det_s_f(?)"                          
#> [107] "det_t"                               
#> [108] "det_t_b(?)"                          
#> [109] "det_t_f(?)"                          
#> [110] "diagnostics.fcast-insample-mean"     
#> [111] "diagnostics.fcast-outsample-mean"    
#> [112] "diagnostics.fcast-outsample-variance"
#> [113] "diagnostics.seas-i-combined"         
#> [114] "diagnostics.seas-i-combined3"        
#> [115] "diagnostics.seas-i-evolutive"        
#> [116] "diagnostics.seas-i-f"                
#> [117] "diagnostics.seas-i-friedman"         
#> [118] "diagnostics.seas-i-kw"               
#> [119] "diagnostics.seas-i-periodogram"      
#> [120] "diagnostics.seas-i-qs"               
#> [121] "diagnostics.seas-i-spectralpeaks"    
#> [122] "diagnostics.seas-i-stable"           
#> [123] "diagnostics.seas-lin-combined"       
#> [124] "diagnostics.seas-lin-evolutive"      
#> [125] "diagnostics.seas-lin-f"              
#> [126] "diagnostics.seas-lin-friedman"       
#> [127] "diagnostics.seas-lin-kw"             
#> [128] "diagnostics.seas-lin-periodogram"    
#> [129] "diagnostics.seas-lin-qs"             
#> [130] "diagnostics.seas-lin-spectralpeaks"  
#> [131] "diagnostics.seas-lin-stable"         
#> [132] "diagnostics.seas-res-combined"       
#> [133] "diagnostics.seas-res-combined3"      
#> [134] "diagnostics.seas-res-evolutive"      
#> [135] "diagnostics.seas-res-f"              
#> [136] "diagnostics.seas-res-friedman"       
#> [137] "diagnostics.seas-res-kw"             
#> [138] "diagnostics.seas-res-periodogram"    
#> [139] "diagnostics.seas-res-qs"             
#> [140] "diagnostics.seas-res-spectralpeaks"  
#> [141] "diagnostics.seas-res-stable"         
#> [142] "diagnostics.seas-sa-ac1"             
#> [143] "diagnostics.seas-sa-combined"        
#> [144] "diagnostics.seas-sa-combined3"       
#> [145] "diagnostics.seas-sa-evolutive"       
#> [146] "diagnostics.seas-sa-f"               
#> [147] "diagnostics.seas-sa-friedman"        
#> [148] "diagnostics.seas-sa-kw"              
#> [149] "diagnostics.seas-sa-periodogram"     
#> [150] "diagnostics.seas-sa-qs"              
#> [151] "diagnostics.seas-sa-spectralpeaks"   
#> [152] "diagnostics.seas-sa-stable"          
#> [153] "diagnostics.seas-si-combined"        
#> [154] "diagnostics.seas-si-combined3"       
#> [155] "diagnostics.seas-si-evolutive"       
#> [156] "diagnostics.seas-si-stable"          
#> [157] "diagnostics.td-i-all"                
#> [158] "diagnostics.td-i-last"               
#> [159] "diagnostics.td-res-all"              
#> [160] "diagnostics.td-res-last"             
#> [161] "diagnostics.td-sa-all"               
#> [162] "diagnostics.td-sa-last"              
#> [163] "ee"                                  
#> [164] "ee_b(?)"                             
#> [165] "ee_f(?)"                             
#> [166] "i"                                   
#> [167] "i_b"                                 
#> [168] "i_eb"                                
#> [169] "i_ef"                                
#> [170] "i_f"                                 
#> [171] "l"                                   
#> [172] "likelihood.adjustedll"               
#> [173] "likelihood.aic"                      
#> [174] "likelihood.aicc"                     
#> [175] "likelihood.bic"                      
#> [176] "likelihood.bic2"                     
#> [177] "likelihood.bicc"                     
#> [178] "likelihood.df"                       
#> [179] "likelihood.hannanquinn"              
#> [180] "likelihood.ll"                       
#> [181] "likelihood.neffectiveobs"            
#> [182] "likelihood.nobs"                     
#> [183] "likelihood.nparams"                  
#> [184] "likelihood.ssqerr"                   
#> [185] "log"                                 
#> [186] "mhe"                                 
#> [187] "mhe_b(?)"                            
#> [188] "mhe_f(?)"                            
#> [189] "mode"                                
#> [190] "omhe"                                
#> [191] "omhe_b(?)"                           
#> [192] "omhe_f(?)"                           
#> [193] "out"                                 
#> [194] "out_b(?)"                            
#> [195] "out_f(?)"                            
#> [196] "out_i"                               
#> [197] "out_i_b(?)"                          
#> [198] "out_i_f(?)"                          
#> [199] "out_s"                               
#> [200] "out_s_b(?)"                          
#> [201] "out_s_f(?)"                          
#> [202] "out_t"                               
#> [203] "out_t_b(?)"                          
#> [204] "out_t_f(?)"                          
#> [205] "period"                              
#> [206] "reg_i"                               
#> [207] "reg_i_b(?)"                          
#> [208] "reg_i_f(?)"                          
#> [209] "reg_s"                               
#> [210] "reg_s_b(?)"                          
#> [211] "reg_s_f(?)"                          
#> [212] "reg_sa"                              
#> [213] "reg_sa_b(?)"                         
#> [214] "reg_sa_f(?)"                         
#> [215] "reg_t"                               
#> [216] "reg_t_b(?)"                          
#> [217] "reg_t_f(?)"                          
#> [218] "reg_u"                               
#> [219] "reg_u_b(?)"                          
#> [220] "reg_u_f(?)"                          
#> [221] "reg_y"                               
#> [222] "reg_y_b(?)"                          
#> [223] "reg_y_f(?)"                          
#> [224] "regression.description"              
#> [225] "regression.details.coefficients"     
#> [226] "regression.details.covar"            
#> [227] "regression.details.covar-ml"         
#> [228] "regression.easter"                   
#> [229] "regression.espan.end"                
#> [230] "regression.espan.missing"            
#> [231] "regression.espan.n"                  
#> [232] "regression.espan.start"              
#> [233] "regression.leaster"                  
#> [234] "regression.lp"                       
#> [235] "regression.mean"                     
#> [236] "regression.missing(*)"               
#> [237] "regression.ml.parameters"            
#> [238] "regression.ml.pcorr"                 
#> [239] "regression.ml.pcovar"                
#> [240] "regression.ml.pcovar-ml"             
#> [241] "regression.ml.pscore"                
#> [242] "regression.mu"                       
#> [243] "regression.nao"                      
#> [244] "regression.nlp"                      
#> [245] "regression.nls"                      
#> [246] "regression.nmh"                      
#> [247] "regression.nout"                     
#> [248] "regression.nso"                      
#> [249] "regression.ntc"                      
#> [250] "regression.ntd"                      
#> [251] "regression.nusers"                   
#> [252] "regression.out(*)"                   
#> [253] "regression.outlier(*)"               
#> [254] "regression.td(*)"                    
#> [255] "regression.type"                     
#> [256] "regression.user(*)"                  
#> [257] "residuals.bp"                        
#> [258] "residuals.bp2"                       
#> [259] "residuals.doornikhansen"             
#> [260] "residuals.kurtosis"                  
#> [261] "residuals.lb"                        
#> [262] "residuals.lb2"                       
#> [263] "residuals.lruns"                     
#> [264] "residuals.ludruns"                   
#> [265] "residuals.mean"                      
#> [266] "residuals.nruns"                     
#> [267] "residuals.nudruns"                   
#> [268] "residuals.res"                       
#> [269] "residuals.seasbp"                    
#> [270] "residuals.seaslb"                    
#> [271] "residuals.ser"                       
#> [272] "residuals.skewness"                  
#> [273] "residuals.tsres"                     
#> [274] "residuals.type"                      
#> [275] "s"                                   
#> [276] "s_b"                                 
#> [277] "s_eb"                                
#> [278] "s_ef"                                
#> [279] "s_f"                                 
#> [280] "sa"                                  
#> [281] "sa_b"                                
#> [282] "sa_eb"                               
#> [283] "sa_ef"                               
#> [284] "sa_f"                                
#> [285] "seasonal"                            
#> [286] "seats.ar_root(*)"                    
#> [287] "seats.ivar-estimate"                 
#> [288] "seats.ivar-estimator"                
#> [289] "seats.ivar-pvalue"                   
#> [290] "seats.ma_root(*)"                    
#> [291] "seats.model_changed"                 
#> [292] "seats.parameters_cutoff"             
#> [293] "seats.savar-estimate"                
#> [294] "seats.savar-estimator"               
#> [295] "seats.savar-pvalue"                  
#> [296] "seats.seasonality"                   
#> [297] "seats.sicorr-estimate"               
#> [298] "seats.sicorr-estimator"              
#> [299] "seats.sicorr-pvalue"                 
#> [300] "seats.svar-estimate"                 
#> [301] "seats.svar-estimator"                
#> [302] "seats.svar-pvalue"                   
#> [303] "seats.ticorr-estimate"               
#> [304] "seats.ticorr-estimator"              
#> [305] "seats.ticorr-pvalue"                 
#> [306] "seats.tscorr-estimate"               
#> [307] "seats.tscorr-estimator"              
#> [308] "seats.tscorr-pvalue"                 
#> [309] "seats.tvar-estimate"                 
#> [310] "seats.tvar-estimator"                
#> [311] "seats.tvar-pvalue"                   
#> [312] "span.end"                            
#> [313] "span.missing"                        
#> [314] "span.n"                              
#> [315] "span.start"                          
#> [316] "t"                                   
#> [317] "t_b"                                 
#> [318] "t_eb"                                
#> [319] "t_ef"                                
#> [320] "t_f"                                 
#> [321] "tde"                                 
#> [322] "tde_b(?)"                            
#> [323] "tde_f(?)"                            
#> [324] "variancedecomposition.cycle"         
#> [325] "variancedecomposition.irregular"     
#> [326] "variancedecomposition.others"        
#> [327] "variancedecomposition.seasonality"   
#> [328] "variancedecomposition.tdh"           
#> [329] "variancedecomposition.total"         
#> [330] "y"                                   
#> [331] "y_b"                                 
#> [332] "y_b(?)"                              
#> [333] "y_eb(?)"                             
#> [334] "y_ef(?)"                             
#> [335] "y_f"                                 
#> [336] "y_f(?)"                              
#> [337] "yc"                                  
#> [338] "ycal"                                
#> [339] "ycal_f(?)"                           
userdefined_variables_tramoseats("tramo")
#>   [1] "adjust"                          "arima.bd"                       
#>   [3] "arima.bp"                        "arima.bphi(*)"                  
#>   [5] "arima.bq"                        "arima.btheta(*)"                
#>   [7] "arima.d"                         "arima.p"                        
#>   [9] "arima.phi(*)"                    "arima.q"                        
#>  [11] "arima.theta(*)"                  "cal"                            
#>  [13] "cal_b(?)"                        "cal_f(?)"                       
#>  [15] "det"                             "det_b(?)"                       
#>  [17] "det_f(?)"                        "det_i"                          
#>  [19] "det_i_b(?)"                      "det_i_f(?)"                     
#>  [21] "det_s"                           "det_s_b(?)"                     
#>  [23] "det_s_f(?)"                      "det_t"                          
#>  [25] "det_t_b(?)"                      "det_t_f(?)"                     
#>  [27] "ee"                              "ee_b(?)"                        
#>  [29] "ee_f(?)"                         "l"                              
#>  [31] "likelihood.adjustedll"           "likelihood.aic"                 
#>  [33] "likelihood.aicc"                 "likelihood.bic"                 
#>  [35] "likelihood.bic2"                 "likelihood.bicc"                
#>  [37] "likelihood.df"                   "likelihood.hannanquinn"         
#>  [39] "likelihood.ll"                   "likelihood.neffectiveobs"       
#>  [41] "likelihood.nobs"                 "likelihood.nparams"             
#>  [43] "likelihood.ssqerr"               "log"                            
#>  [45] "mhe"                             "mhe_b(?)"                       
#>  [47] "mhe_f(?)"                        "omhe"                           
#>  [49] "omhe_b(?)"                       "omhe_f(?)"                      
#>  [51] "out"                             "out_b(?)"                       
#>  [53] "out_f(?)"                        "out_i"                          
#>  [55] "out_i_b(?)"                      "out_i_f(?)"                     
#>  [57] "out_s"                           "out_s_b(?)"                     
#>  [59] "out_s_f(?)"                      "out_t"                          
#>  [61] "out_t_b(?)"                      "out_t_f(?)"                     
#>  [63] "period"                          "reg_i"                          
#>  [65] "reg_i_b(?)"                      "reg_i_f(?)"                     
#>  [67] "reg_s"                           "reg_s_b(?)"                     
#>  [69] "reg_s_f(?)"                      "reg_sa"                         
#>  [71] "reg_sa_b(?)"                     "reg_sa_f(?)"                    
#>  [73] "reg_t"                           "reg_t_b(?)"                     
#>  [75] "reg_t_f(?)"                      "reg_u"                          
#>  [77] "reg_u_b(?)"                      "reg_u_f(?)"                     
#>  [79] "reg_y"                           "reg_y_b(?)"                     
#>  [81] "reg_y_f(?)"                      "regression.description"         
#>  [83] "regression.details.coefficients" "regression.details.covar"       
#>  [85] "regression.details.covar-ml"     "regression.easter"              
#>  [87] "regression.espan.end"            "regression.espan.missing"       
#>  [89] "regression.espan.n"              "regression.espan.start"         
#>  [91] "regression.leaster"              "regression.lp"                  
#>  [93] "regression.mean"                 "regression.missing(*)"          
#>  [95] "regression.ml.parameters"        "regression.ml.pcorr"            
#>  [97] "regression.ml.pcovar"            "regression.ml.pcovar-ml"        
#>  [99] "regression.ml.pscore"            "regression.mu"                  
#> [101] "regression.nao"                  "regression.nlp"                 
#> [103] "regression.nls"                  "regression.nmh"                 
#> [105] "regression.nout"                 "regression.nso"                 
#> [107] "regression.ntc"                  "regression.ntd"                 
#> [109] "regression.nusers"               "regression.out(*)"              
#> [111] "regression.outlier(*)"           "regression.td(*)"               
#> [113] "regression.type"                 "regression.user(*)"             
#> [115] "residuals.bp"                    "residuals.bp2"                  
#> [117] "residuals.doornikhansen"         "residuals.kurtosis"             
#> [119] "residuals.lb"                    "residuals.lb2"                  
#> [121] "residuals.lruns"                 "residuals.ludruns"              
#> [123] "residuals.mean"                  "residuals.nruns"                
#> [125] "residuals.nudruns"               "residuals.res"                  
#> [127] "residuals.seasbp"                "residuals.seaslb"               
#> [129] "residuals.ser"                   "residuals.skewness"             
#> [131] "residuals.tsres"                 "residuals.type"                 
#> [133] "span.end"                        "span.missing"                   
#> [135] "span.n"                          "span.start"                     
#> [137] "tde"                             "tde_b(?)"                       
#> [139] "tde_f(?)"                        "y"                              
#> [141] "y_b(?)"                          "y_eb(?)"                        
#> [143] "y_ef(?)"                         "y_f(?)"                         
#> [145] "yc"                              "ycal"                           
#> [147] "ycal_f(?)"                      
y = rjd3toolkit::ABS$X0.2.09.10.M
m <- tramoseats(y,"rsafull", userdefined=c("ycal","variancedecomposition.seasonality"))
m$user_defined$ycal
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1982                                443.8278  510.6464  441.5556  447.2441
#> 1983  385.2708  381.9008  463.5100  496.9624  524.2277  456.2137  471.7869
#> 1984  425.7682  401.8572  484.2305  490.2518  587.0210  494.0008  557.9031
#> 1985  506.9057  457.1696  543.2305  576.5562  692.6467  564.5811  606.3516
#> 1986  564.2652  483.1348  542.3265  606.3810  732.1905  570.9575  651.1328
#> 1987  598.0465  518.4960  631.6257  648.5345  719.3164  674.5692  713.0216
#> 1988  635.6984  540.2656  665.7617  706.5474  760.5717  712.3406  688.5893
#> 1989  648.3545  557.6964  693.4493  718.9404  769.3060  777.8438  726.5485
#> 1990  682.1958  592.2493  726.0960  748.6074  800.6617  821.7550  729.2576
#> 1991  681.3724  610.9402  706.5284  771.9465  819.1489  750.5975  799.1810
#> 1992  731.9128  635.8665  738.5989  799.8688  885.1420  831.0129  779.5861
#> 1993  767.7745  649.6355  777.8968  810.9673  921.1143  882.6137  751.5766
#> 1994  757.6383  671.2563  806.5525  821.0640  932.2342  900.3907  788.7294
#> 1995  772.4813  689.5431  822.8058  887.8325  932.6628  897.0479  894.6980
#> 1996  838.0084  705.3053  838.9181  878.2254  952.2779  913.0980  913.3497
#> 1997  831.7101  734.9064  849.9486  866.2586  969.1100  856.5394  953.4305
#> 1998  893.6114  723.5909  866.4191  933.3085  977.4240  854.5549 1030.9091
#> 1999  949.0141  746.0200  916.1835  935.0630 1027.3889  889.8768 1027.2683
#> 2000  939.7073  770.2115  914.8035 1004.1523 1058.4999 1152.9657  873.0570
#> 2001  977.6262  823.0063 1016.5144 1045.4950 1103.0575 1011.7933 1130.1799
#> 2002 1009.1970  850.0829 1025.5075 1109.4281 1177.8851 1148.6583 1082.0378
#> 2003 1087.2862  909.7928 1138.5197 1185.9589 1205.1528 1277.1743 1160.4931
#> 2004 1152.3525  979.4507 1167.1031 1231.9334 1300.7971 1377.1002 1272.1761
#> 2005 1266.2527 1048.0045 1245.3569 1168.0058 1297.9410 1434.7448 1299.7060
#> 2006 1225.5594 1048.4086 1221.7328 1321.5573 1275.0684 1413.6319 1434.6055
#> 2007 1274.7666 1057.8046 1330.3101 1310.4905 1340.2615 1472.4263 1501.3282
#> 2008 1372.3665 1108.6542 1339.1719 1424.9903 1384.4011 1446.0679 1553.5095
#> 2009 1413.5290 1067.5036 1460.2463 1451.8667 1530.7133 1509.6064 1553.5344
#> 2010 1480.0388 1123.3743 1399.7360 1363.2575 1450.0489 1543.5542 1540.1718
#> 2011 1435.2151 1129.0321 1330.6594 1417.5987 1446.5030 1452.2080 1543.9109
#> 2012 1398.7619 1097.1817 1347.8930 1399.6272 1430.3912 1595.7290 1461.3900
#> 2013 1371.9738 1125.0919 1372.6810 1381.6006 1426.1420 1545.0108 1398.7931
#> 2014 1435.1426 1075.8893 1347.4271 1399.8856 1375.3413 1507.9715 1474.2779
#> 2015 1433.2073 1064.6748 1427.4659 1399.2055 1439.3380 1489.8874 1524.0601
#> 2016 1530.5119 1133.0774 1389.6337 1468.9923 1488.3919 1574.7464 1479.2329
#> 2017 1466.6261 1103.6731 1389.9049 1492.0682 1460.9572 1540.4532 1468.6419
#>            Aug       Sep       Oct       Nov       Dec
#> 1982  450.1007  458.8776  456.0708  526.0734  879.5812
#> 1983  460.8996  477.5088  493.0695  565.8239  938.8132
#> 1984  497.8985  525.3881  559.8794  616.4978 1056.0404
#> 1985  567.4574  580.9652  625.5077  688.1853 1219.1939
#> 1986  585.4268  665.9170  658.3259  704.3698 1375.9594
#> 1987  611.2314  668.9000  690.7880  766.7792 1433.5330
#> 1988  654.7350  715.5213  741.2806  865.8987 1529.4559
#> 1989  710.6302  748.2078  771.7626  911.4408 1581.0856
#> 1990  751.1008  744.7688  791.2345  921.4822 1588.1246
#> 1991  750.5050  765.0252  842.9775  932.6189 1643.0114
#> 1992  735.1845  815.6541  879.1936  999.3303 1620.8613
#> 1993  748.1487  833.3991  893.5038  975.8868 1691.8066
#> 1994  803.5063  877.4608  949.5626  995.3407 1770.5644
#> 1995  857.9991  845.8430  945.1704 1053.1197 1871.0286
#> 1996  865.9448  860.1504  916.7110 1067.7682 1886.6448
#> 1997  842.2248  944.7993  966.2238 1116.8014 1890.1712
#> 1998  883.8267  934.6492  994.9257 1173.7953 1939.7515
#> 1999  923.5073  988.8906 1051.4713 1214.5264 2046.9821
#> 2000  990.8373 1041.1135 1098.9686 1216.0109 2193.7135
#> 2001  952.4757 1017.9143 1138.6695 1330.3374 2305.8338
#> 2002 1038.3739 1093.7333 1189.3582 1475.9803 2401.0163
#> 2003 1109.3995 1205.4718 1301.7168 1535.7652 2438.4828
#> 2004 1197.9414 1300.2683 1368.4138 1546.0261 2473.4697
#> 2005 1217.8332 1264.0580 1335.4430 1532.6098 2470.5097
#> 2006 1216.6523 1203.5098 1380.4868 1601.5825 2630.5421
#> 2007 1301.1234 1359.9994 1449.0877 1669.7476 2790.9783
#> 2008 1277.8444 1391.6951 1436.5273 1684.1964 2903.1042
#> 2009 1362.9769 1392.8203 1494.6778 1791.6963 2744.2422
#> 2010 1359.3370 1402.9756 1475.8075 1715.8505 2722.3201
#> 2011 1301.1210 1331.2277 1427.1887 1651.6050 2659.9868
#> 2012 1344.1480 1374.5712 1451.0995 1654.9089 2797.3792
#> 2013 1302.6678 1402.5291 1436.6748 1647.5577 2798.2425
#> 2014 1263.5387 1364.6319 1462.5398 1696.0548 2773.1430
#> 2015 1353.2232 1393.0193 1525.8515 1785.4028 2860.5860
#> 2016 1301.5234 1378.5137 1521.3728 1675.7821 2776.7905
#> 2017 1279.3896                                        
m$user_defined$variancedecomposition.seasonality
#> [1] 0.9704556